# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API for managing mechanics, customers, inventory, and service tickets."
  version: "1.0.0"
host: "127.0.0.1:5000" # render name: "your-app-name.onrender.com"
schemes:
  - "https"
basePath: "/"
securityDefinitions:
  BearerAuth:
    type: apiKey
    swagger: "2.0"
    in: header
    description: "JWT token using Bearer token number. Example: 'Bearer {token number}'"
paths:
...
  /service-tickets/customer/my-tickets:
    get:
      tags:
        - ServiceTicket
      summary: Customer gets own tickets (paginated)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of own tickets
        '500':
          description: Server error
...
  /customers/:
    post:
      tags: [Customers]
      summary: "Create customer (admin only)"
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerRegister"
      responses:
        201:
          description: "Customer created"
        400:
          description: "Invalid input"
        409:
          description: "Email already exists"
        500:
          description: "Internal server error"
  /customers/login:
    post:
      tags: [Customers]
      summary: "Customer login"
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            $ref: "#/definitions/CustomerLogin"
      responses:
        200:
          description: "Login successful (JWT token returned)"
        401:
          description: "Invalid credentials"
        500:
          description: "Internal server error"
  /customers/register:
    post:
      tags: [Customers]
      summary: "Register new customer (public)"
      parameters:
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerRegister"
      responses:
        201:
          description: "Customer registered"
        400:
          description: "Invalid input"
        409:
          description: "Email already exists"
        500:
          description: "Internal server error"
  /customers/my-tickets:
    get:
      tags: [Customers]
      summary: "Get tickets for current customer"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "List of tickets"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /customers/{id}:
    get:
      tags: [Customers]
      summary: "Get own customer profile"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Customer profile"
        403:
          description: "Forbidden"
        404:
          description: "Customer not found"
        500:
          description: "Internal server error"
    put:
      tags: [Customers]
      summary: "Update own customer profile"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerRegister"
      responses:
        200:
          description: "Customer updated"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Customer not found"
        500:
          description: "Internal server error"
  /customers/admin/{id}:
    get:
      tags: [Customers]
      summary: "Admin gets customer by id"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Customer profile"
        404:
          description: "Customer not found"
        500:
          description: "Internal server error"
  /customers/admin/update/{id}:
    put:
      tags: [Customers]
      summary: "Admin updates customer"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerRegister"
      responses:
        200:
          description: "Customer updated"
        400:
          description: "Invalid input"
        404:
          description: "Customer not found"
        500:
          description: "Internal server error"
  /customers/admin/delete/{id}:
    delete:
      tags: [Customers]
      summary: "Admin deletes customer"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Customer deleted"
        404:
          description: "Customer not found"
        500:
          description: "Internal server error"
definitions:
...
  CustomerLogin:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
  CustomerRegister:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      password:
        type: string
    required:
      - name
      - email
      - phone
      - password
...